requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.5
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
model-bakery==1.3.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.3.0
